+++ lib/tankard/client.rb
@@ -1,6 +1,7 @@
 require 'tankard/request'
 require 'tankard/api/beer'
 require 'tankard/api/beers'
+require 'tankard/api/styles'
 
 module Tankard
   class Client
@@ -19,6 +20,10 @@ module Tankard
 
     def beers(options={})
       Tankard::Api::Beers.new(@tankard_request, options)
+    end
+
+    def styles
+      Tankard::Api::Styles.new(@tankard_request)
     end
   end
 end