+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.1.0 (unreleased)
 * Add params method to beers.  Allows passing additional parameters in an additional way.
 * Add params method to beer.  Allows passing additional parameters in an additional way.
+* Add styles method & class.  Allows querying of the /styles route
 
 ## v0.0.1
 * Support for querying a beer